% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{alpha_1}
\alias{alpha_1}
\title{Vector that defines the success probability null curve.}
\format{
a length 2 vector
\describe{
The matrix multiplication of this vector with g_t_1 defines the MEE under the
null hypothesis.
}
}
\usage{
alpha_1
}
\description{
Vector that defines the success probability null curve.
}
\keyword{datasets}
