% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{beta_1}
\alias{beta_1}
\title{Vector that defines the MEE under the alternative hypothesis.}
\format{
a length 2 vector
\describe{
The matrix multiplication of this vector with f_t_1 defines the MEE under the
alternative hypothesis.
}
}
\usage{
beta_1
}
\description{
Vector that defines the MEE under the alternative hypothesis.
}
\keyword{datasets}
