% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mrt_bin_samplesize_f.R
\name{calculate_mrt_bin_samplesize_f}
\alias{calculate_mrt_bin_samplesize_f}
\title{Returns sample size needed to achieve a specified power in the context of a
binary MRT.}
\usage{
calculate_mrt_bin_samplesize_f(
  avail_pattern,
  f_t,
  g_t,
  beta,
  alpha,
  p_t,
  gamma,
  b,
  exact = FALSE
)
}
\arguments{
\item{avail_pattern}{A vector of length T that is the average availability at
each time point}

\item{f_t}{Defines marginal excursion effect MEE(t) under
alternative together with beta}

\item{g_t}{Defines success probability null curve together with
alpha}

\item{beta}{Defines marginal excursion effect MEE(t) under
alternative together with g_t}

\item{alpha}{Defines success probability null curve together with f_t}

\item{p_t}{Randomization probability at each time point}

\item{gamma}{Desired Type I error}

\item{b}{Desired Type II error}

\item{exact}{Determines if exact n or ceiling will be returned}
}
\value{
Sample size to achieve desired power.
}
\description{
Returns sample size needed to achieve a specified power in the context of a
binary MRT.
}
