% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_m_sigma.R
\name{compute_m_sigma}
\alias{compute_m_sigma}
\title{Computes "M" and "Sigma" matrices for the sandwich estimator of
variance-covariance matrix. Primarily a helper function for
calculate_mrt_bin_power_f and calculate_mrt_bin_samplesize_f.}
\usage{
compute_m_sigma(avail_pattern, f_t, g_t, beta, alpha, p_t)
}
\arguments{
\item{avail_pattern}{A vector of length T that is the average availability at
each time point}

\item{f_t}{Defines marginal excursion effect MEE(t) under
alternative together with beta}

\item{g_t}{Defines success probability null curve together with
alpha}

\item{beta}{Defines marginal excursion effect MEE(t) under
alternative together with g_t}

\item{alpha}{Defines success probability null curve together with f_t}

\item{p_t}{Randomization probability at each time point}
}
\value{
List containing two matrices. The first is the M matrix
  and the second is the Sigma matrix.
}
\description{
Computes "M" and "Sigma" matrices for the sandwich estimator of
variance-covariance matrix. Primarily a helper function for
calculate_mrt_bin_power_f and calculate_mrt_bin_samplesize_f.
}
