% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_samp.R
\name{max_samp}
\alias{max_samp}
\title{Returns default maximum sample size to end power_vs_n_plot().}
\usage{
max_samp(min_samp)
}
\arguments{
\item{min_samp}{The starting sample size of the plot.}
}
\value{
A default maximum sample size to end power_vs_n_plot().
}
\description{
Returns default maximum sample size to end power_vs_n_plot().
}
\examples{
max_samp(100)
}
