% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_samp.R
\name{min_samp}
\alias{min_samp}
\title{Compute minimum sample size.}
\usage{
min_samp(alph, bet)
}
\arguments{
\item{alph}{Vector to describe the MEE under the alternative.}

\item{bet}{Vector to describe the MEE under the null.}
}
\value{
A default minimum sample size to start power_vs_n_plot() at.
}
\description{
Returns a default minimum sample size to start power_vs_n_plot() at.
}
\examples{
min_samp(alpha_1, beta_1)
}
