% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt_binary_ss.R
\name{mrt_binary_ss}
\alias{mrt_binary_ss}
\title{Returns sample size needed to achieve a specified power in the context of an
MRT with binary outcomes.}
\usage{
mrt_binary_ss(
  avail_pattern,
  f_t,
  g_t,
  beta,
  alpha,
  p_t,
  gamma,
  b,
  exact = FALSE
)
}
\arguments{
\item{avail_pattern}{A vector of length m that is the average availability at
each time point}

\item{f_t}{Defines marginal excursion effect MEE(t) under
alternative together with beta. Assumed to be matrix of size T*p.}

\item{g_t}{Defines success probability null curve together with
alpha. Assumed to be matrix of size T*q.}

\item{beta}{Length p vector that defines marginal excursion effect
MEE(t) under alternative together with f_t.}

\item{alpha}{Length q vector that defines success probability null
curve together with g_t.}

\item{p_t}{Length m vector of Randomization probabilities at each
time point.}

\item{gamma}{Desired Type I error}

\item{b}{Desired Type II error}

\item{exact}{Determines if exact n or ceiling will be returned}
}
\value{
Sample size to achieve desired power.
}
\description{
Returns sample size needed to achieve a specified power in the context of an
MRT with binary outcomes.
}
\examples{
           mrt_binary_ss(tau_t_1, f_t_1, g_t_1, 
                                         beta_1, alpha_1, p_t_1, 
                                         0.05, .2, FALSE)
}
