% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{p_t_1}
\alias{p_t_1}
\title{A vector of probabilities for each time point.}
\format{
a length T vector
\describe{
vectro of probs
}
}
\usage{
p_t_1
}
\description{
A vector of probabilities for each time point.
}
\keyword{datasets}
