% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-vs-n-plot.R
\name{plot_power_vs_samplesize}
\alias{plot_power_vs_samplesize}
\title{Returns a plot of power vs sample size in the context of MRT}
\usage{
plot_power_vs_samplesize(avail_pattern, f_t, g_t, beta, alpha, p_t, gamma, n)
}
\arguments{
\item{avail_pattern}{A vector of length T that is the average availability at
each time point}

\item{f_t}{Defines marginal excursion effect MEE(t) under
alternative together with beta}

\item{g_t}{Defines success probability null curve together with
alpha}

\item{beta}{Defines marginal excursion effect MEE(t) under
alternative together with g_t}

\item{alpha}{Defines success probability null curve together with f_t}

\item{p_t}{Randomization probability at each time point}

\item{gamma}{Desired Type I error}

\item{n}{Number of participants, sample size}
}
\value{
Plot of power and sample sice
}
\description{
Returns a plot of power vs sample size in the context of MRT
}
