% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_summary.R
\name{power_summary}
\alias{power_summary}
\title{Reports the power of a binary outcome MRT at a range of power levels.}
\usage{
power_summary(
  avail_pattern,
  f_t,
  g_t,
  beta,
  alpha,
  p_t,
  gamma,
  power_levels = seq(from = 0.6, to = 0.95, by = 0.05)
)
}
\arguments{
\item{avail_pattern}{A vector of length T that is the average availability at
each time point}

\item{f_t}{Defines marginal excursion effect MEE(t) under
alternative together with beta. Assumed to be matrix of size T*p.}

\item{g_t}{Defines success probability null curve together with
alpha. Assumed to be matrix of size T*q.}

\item{beta}{Length p vector that defines marginal excursion effect
MEE(t) under alternative together with g_t.}

\item{alpha}{Length q vector that defines success probability null
curve together with f_t.}

\item{p_t}{Length T vector of Randomization probabilities at each
time point.}

\item{gamma}{Desired Type I error}

\item{power_levels}{Vector of powers to find sample size for.}
}
\value{
Table containing needed sample size to achieve some
  user-specified power values.
}
\description{
Reports the power of a binary outcome MRT at a range of power levels.
}
\examples{
           power_summary(tau_t_1, f_t_1, g_t_1, 
beta_1, alpha_1, p_t_1, gamma_1)
}
