% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-vs-n-plot.R
\name{power_vs_n_plot}
\alias{power_vs_n_plot}
\title{Returns a plot of power vs sample size in the context of MRT}
\usage{
power_vs_n_plot(avail_pattern, f_t, g_t, beta, alpha, p_t, gamma)
}
\arguments{
\item{avail_pattern}{A vector of length T that is the average availability at
each time point}

\item{f_t}{Defines marginal excursion effect MEE(t) under
alternative together with beta}

\item{g_t}{Defines success probability null curve together with
alpha}

\item{beta}{Defines marginal excursion effect MEE(t) under
alternative together with g_t}

\item{alpha}{Defines success probability null curve together with f_t}

\item{p_t}{Randomization probability at each time point}

\item{gamma}{Desired Type I error}
}
\value{
Plot of power and sample size
}
\description{
Returns a plot of power vs sample size in the context of MRT
}
\examples{
           power_vs_n_plot(tau_t_1, f_t_1, g_t_1, beta_1, alpha_1,
                        p_t_1, 0.05)
}
