% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{sigma_matrix_1}
\alias{sigma_matrix_1}
\title{An example matrix for "meat" of sandwich estimator of variance.}
\format{
A 2 by 2 matrix
\describe{
Generated from a toy example.
}
}
\usage{
sigma_matrix_1
}
\description{
An example matrix for "meat" of sandwich estimator of variance.
}
\keyword{datasets}
