% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{tau_t_1}
\alias{tau_t_1}
\title{Time avail vec}
\format{
vector of length T
\describe{A vector of length T that is the aver availability at each time point.}
}
\usage{
tau_t_1
}
\description{
Time avail vec
}
\keyword{datasets}
