% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mrt_bin_power_f.R
\name{calculate_mrt_bin_power_f}
\alias{calculate_mrt_bin_power_f}
\title{Returns power of a test given a prespecified sample size in the context of a
binary MRT.}
\usage{
calculate_mrt_bin_power_f(avail_pattern, f_t, g_t, beta, alpha, p_t, gamma, n)
}
\arguments{
\item{avail_pattern}{A vector of length T that is the average availability at
each time point}

\item{f_t}{Defines marginal excursion effect MEE(t) under
alternative together with beta}

\item{g_t}{Defines success probability null curve together with
alpha}

\item{beta}{Defines marginal excursion effect MEE(t) under
alternative together with g_t}

\item{alpha}{Defines success probability null curve together with f_t}

\item{p_t}{Randomization probability at each time point}

\item{gamma}{Desired Type I error}

\item{n}{Number of participants}
}
\value{
Power of the test.

Power of the test for fixed null/alternative and sample
  size.
}
\description{
Returns power of a test given a prespecified sample size in the context of a
binary MRT.
}
\examples{
           calculate_mrt_bin_power_f(tau_t_1, f_t_1, g_t_1, beta_1, 
alpha_1, p_t_1, 0.05, 100)
}
