% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ncp.R
\name{compute_ncp}
\alias{compute_ncp}
\title{Computes the non-centrality parameter for an F distributed random variable
in the context of a MRT with binary outcome.}
\usage{
compute_ncp(x, beta, m_matrix, sigma_matrix)
}
\arguments{
\item{x}{Sample size}

\item{beta}{Marginal excursion effect, assumed dimension p}

\item{m_matrix}{"Bread" of sandwich estimator for variance}

\item{sigma_matrix}{"Meat" of sandwich estimator for variance}
}
\value{
Returns non-centrality parameter for an F distributed random
  variable.
}
\description{
A helper function for
mrt_binary_power() and mrt_binary_ss().
}
\examples{
compute_ncp(300, beta_1, m_matrix_1, sigma_matrix_1)
}
