% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{f_t_1}
\alias{f_t_1}
\title{A matrix defining the MEE under the alternative hypothesis.}
\format{
a 10 by 2 matrix
\describe{
In this example it is a log-linear trend.
}
}
\usage{
f_t_1
}
\description{
A matrix defining the MEE under the alternative hypothesis.
}
\keyword{datasets}
