% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{g_t_1}
\alias{g_t_1}
\title{A matrix defining the success probability null curve.}
\format{
a 10 by 2 matrix
\describe{
In this example it is a log-linear trend.
}
}
\usage{
g_t_1
}
\description{
A matrix defining the success probability null curve.
}
\keyword{datasets}
