% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_full_column_rank.R
\name{is_full_column_rank}
\alias{is_full_column_rank}
\title{Check if a matrix is full column rank.}
\usage{
is_full_column_rank(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
Boolean TRUE/FALSE for if matrix is full column rank.
}
\description{
Used in checking if p_t*f_t is in the linear span of g_t.
}
\examples{
  is_full_column_rank(diag(4))
}
