% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{p_t_1}
\alias{p_t_1}
\title{A vector of randomization probabilities for each time point.}
\format{
a length T vector
\describe{
Vector of randomization probabilities. 
}
}
\usage{
p_t_1
}
\description{
A vector of randomization probabilities for each time point.
}
\keyword{datasets}
