% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_summary.R
\name{power_summary}
\alias{power_summary}
\title{Title This function returns a table of selected sample sizes based on power values}
\usage{
power_summary(avail_pattern, f_t, g_t, beta, alpha, p_t, gamma)
}
\arguments{
\item{avail_pattern}{A vector of length T that is the average availability at
each time point}

\item{f_t}{Defines marginal excursion effect MEE(t) under
alternative together with beta}

\item{g_t}{Defines success probability null curve together with
alpha}

\item{beta}{Defines marginal excursion effect MEE(t) under
alternative together with g_t}

\item{alpha}{Defines success probability null curve together with f_t}

\item{p_t}{Randomization probability at each time point}

\item{gamma}{Desired Type I error}
}
\value{
Table of the input values, and sample sizes with some selected power values
}
\description{
Title This function returns a table of selected sample sizes based on power values
}
\examples{
  power_summary(tau_t_1, f_t_1, g_t_1, beta_1, alpha_1, p_t_1, gamma_1)
}
