% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{tau_t_1}
\alias{tau_t_1}
\title{Vector that holds the average availability at each time point.}
\format{
vector of length T
\describe{A vector of length T that is the average availability
 at each decision point.}
}
\usage{
tau_t_1
}
\description{
Vector that holds the average availability at each time point.
}
\keyword{datasets}
